Melhorado tipagem dos dados na database

+++ codigo/public/assets/js/components/create-entrie-popup.js
@@ -23,7 +23,7 @@ export function createEntriePopup() {
       categoryId: parseInt(categoryId.value) || undefined,
       label: label.value,
       type: type.checked ? "income" : "expense",
-      value: parseInt(value.value),
+      value: parseFloat(value.value),
       date: new Date(date.value).getTime(),
     };
